requirements.txt
Django==1.4.3
dogstatsd-python==0.2
path.py==3.0
pip
scipy==0.11.0
South==0.7.6
reportlab==2.6
requests==1.1.0
pytz==2012h
matplotlib==1.2.0
MySQL-python==1.2.4
boto==2.6.0
lxml==3.0.1
gunicorn
wsgiref
-e git+https://github.com/edx/ease.git@8cc465cbcca2f925f52581654763fde618cd1203#egg=ease
django-celery==3.0.11
celery-with-redis==3.0
